Move types' and interfaces' declarations into separate files
