+++ lib/methods.d.ts
