+++ lib/methods.js
