+++ src/auxiliary.ts
@@ -1,0 +1,4 @@
+export interface Fraction {
+  num: number;
+  den: number;
+}
